-- add Authorize button in Swagger -- fix NaN in INCOME transactionDto -- add field 'isMandatory' to transaction

+++ src/main/kotlin/main/CategoryController.kt
@@ -1,11 +1,15 @@
 package main
 
 import entities.Category
+import exceptions.UserNotAuthorizedException
+import io.swagger.v3.oas.annotations.security.SecurityRequirement
+import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.web.bind.annotation.*
 import java.util.logging.Logger
 
 @RestController
+@Tag(name = "Categories")
 class CategoryController(
     @Autowired
     val managerBeans: ManagerBeans
@@ -28,13 +32,23 @@ class CategoryController(
     }
 
     @PostMapping("/categories/insert")
-    fun addCategory(@RequestBody category: Category, @RequestHeader("Authorization") auth: String): Category {
+    @SecurityRequirement(name = "basicAuth")
+    fun addCategory(@RequestBody category: Category, @RequestHeader("Authorization") auth: String?): Category {
+        if (auth == null) {
+            log.severe("Authorization required")
+            throw UserNotAuthorizedException("Authorization required")
+        }
         log.info("Category to add: $category")
         return categoryService.addCategory(category, auth)
     }
 
     @DeleteMapping("/categories/delete/{id}")
-    fun deleteCategory(@PathVariable id: Long, @RequestHeader("Authorization") auth: String): String {
+    @SecurityRequirement(name = "basicAuth")
+    fun deleteCategory(@PathVariable id: Long, @RequestHeader("Authorization") auth: String?): String {
+        if (auth == null) {
+            log.severe("Authorization required")
+            throw UserNotAuthorizedException("Authorization required")
+        }
         log.info("Category id to delete: $id")
         return categoryService.deleteCategory(id, auth)
     }
